Move tooltip functionality to new class SQLHoverTooltip
