+++ src/query.ts
@@ -25,7 +25,7 @@ import { LogLevels, logMessage } from './logger';
 import * as token from './tokens';
 import { IEnvResponse } from './environment';
 import { apiRequest } from './request';
-import { format as formatSQL } from 'sql-formatter';
+import { SQLHoverTooltip } from './sql-tooltip';
 
 /**
  * The command IDs used by the plugin.
@@ -324,7 +324,8 @@ async function getRecentQueryMenu(
     });
 
     // Add single event delegation for all menu items
-    addHoverTooltipsToMenu(retval, queryDataMap);
+    const sqlTooltip = new SQLHoverTooltip(queryDataMap);
+    sqlTooltip.attachToMenu(retval);
   } catch (error) {
     logMessage(
       LogLevels.ERROR,
@@ -443,515 +444,11 @@ async function openQueryFromJobref(
 }
 
 /**
- * Add hover tooltip functionality to all menu items using event delegation
- */
-function addHoverTooltipsToMenu(
-  menu: Menu,
-  queryDataMap: Map<string, { sqlText: string; jobref: string }>
-): void {
-  const menuNode = menu.node;
-  if (!menuNode) {
-    return;
-  }
-
-  // Track current hovered menu item to prevent rapid toggling
-  let currentHoveredJobref: string | null = null;
-
-  // Single event delegation for all menu items
-  const handleMouseEnter = (event: MouseEvent) => {
-    const target = event.target as HTMLElement;
-    if (!target) {
-      return;
-    }
-
-    // Find the closest menu item element
-    const menuItem = target.closest('[data-command^="q-"]') as HTMLElement;
-
-    if (!menuItem) {
-      return;
-    }
-
-    const commandAttr = menuItem.getAttribute('data-command');
-
-    if (!commandAttr || !commandAttr.startsWith('q-')) {
-      return;
-    }
-
-    const jobref = commandAttr.substring(2); // Remove 'q-' prefix
-
-    const queryData = queryDataMap.get(jobref);
-
-    if (!queryData) {
-      return;
-    }
-
-    // If we're already hovering over this item, don't show tooltip again
-    if (currentHoveredJobref === jobref) {
-      return;
-    }
-
-    // Clear any pending hide timeout when entering new item
-    if (tooltipHideTimeout) {
-      clearTimeout(tooltipHideTimeout);
-      tooltipHideTimeout = null;
-    }
-
-    // Clear any pending show timeout from previous item
-    if (tooltipShowTimeout) {
-      clearTimeout(tooltipShowTimeout);
-      tooltipShowTimeout = null;
-    }
-
-    // Update current hovered item
-    currentHoveredJobref = jobref;
-
-    // Add a small delay before showing tooltip to prevent rapid toggling
-    tooltipShowTimeout = window.setTimeout(() => {
-      showSQLTooltip(event, queryData.sqlText, queryData.jobref, () => {
-        currentHoveredJobref = null;
-      });
-    }, 150); // Small delay to prevent flashy behavior
-  };
-
-  const handleMouseLeave = (event: MouseEvent) => {
-    const target = event.target as HTMLElement;
-    if (!target) {
-      return;
-    }
-
-    const menuItem = target.closest('[data-command^="q-"]') as HTMLElement;
-
-    if (!menuItem) {
-      return;
-    }
-
-    const commandAttr = menuItem.getAttribute('data-command');
-
-    if (!commandAttr || !commandAttr.startsWith('q-')) {
-      return;
-    }
-
-    const jobref = commandAttr.substring(2);
-
-    // Only process if we're leaving the item we're currently tracking
-    if (currentHoveredJobref !== jobref) {
-      return;
-    }
-
-    // DON'T clear the show timeout here - let it complete if user was hovering long enough
-    // Only set the hide timeout
-
-    // Clear any existing hide timeout before setting a new one
-    if (tooltipHideTimeout) {
-      clearTimeout(tooltipHideTimeout);
-      tooltipHideTimeout = null;
-    }
-
-    // Add a longer delay before hiding to allow mouse to move to tooltip
-    tooltipHideTimeout = window.setTimeout(() => {
-      hideSQLTooltip();
-      currentHoveredJobref = null;
-    }, 300); // Longer delay to allow mouse movement to tooltip
-  };
-
-  const handleClick = (event: MouseEvent) => {
-    const target = event.target as HTMLElement;
-    if (!target) {
-      return;
-    }
-
-    const menuItem = target.closest('[data-command^="q-"]') as HTMLElement;
-
-    if (!menuItem) {
-      return;
-    }
-
-    const commandAttr = menuItem.getAttribute('data-command');
-
-    if (!commandAttr || !commandAttr.startsWith('q-')) {
-      return;
-    }
-
-    hideSQLTooltip();
-    currentHoveredJobref = null;
-  };
-
-  // Add event listeners with capture to ensure they're processed
-  menuNode.addEventListener('mouseenter', handleMouseEnter, true);
-  menuNode.addEventListener('mouseleave', handleMouseLeave, true);
-  menuNode.addEventListener('click', handleClick, true);
-}
-
-/**
- * Global tooltip element
- */
-let globalTooltip: HTMLElement | null = null;
-
-/**
- * Global tooltip hide timeout
- */
-let tooltipHideTimeout: number | null = null;
-
-/**
- * Global tooltip show timeout
- */
-let tooltipShowTimeout: number | null = null;
-
-/**
- * Show SQL tooltip on hover with syntax highlighting
- */
-function showSQLTooltip(
-  event: MouseEvent,
-  sqlText: string,
-  jobref: string,
-  resetHoveredJobref?: () => void
-): void {
-  // Remove existing tooltip only if one exists
-  if (globalTooltip) {
-    hideSQLTooltip();
-  }
-
-  // Create tooltip element
-  globalTooltip = document.createElement('div');
-  globalTooltip.className = 'sql-hover-tooltip';
-  globalTooltip.style.cssText = `
-    position: fixed;
-    z-index: 10000;
-    background: #ffffff;
-    border: 1px solid #e1e4e8;
-    border-radius: 8px;
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-    padding: 16px;
-    min-width: 500px;
-    max-width: 700px;
-    min-height: 200px;
-    max-height: 500px;
-    overflow: hidden;
-    font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-    font-size: 12px;
-    line-height: 1.4;
-    cursor: text;
-    user-select: text;
-  `;
-
-  // Create title
-  const title = document.createElement('div');
-  title.textContent = `Query: ${jobref}`;
-  title.style.cssText = `
-    font-weight: 600;
-    color: #24292e;
-    margin-bottom: 8px;
-    font-size: 13px;
-  `;
-
-  // Create container for SQL display
-  const sqlContainer = document.createElement('div');
-  sqlContainer.style.cssText = `
-    margin: 0;
-    background: #f6f8fa;
-    border: 1px solid #e1e4e8;
-    border-radius: 4px;
-    overflow: auto;
-    max-height: 450px;
-  `;
-
-  // Create pre element for SQL with syntax highlighting
-  const sqlPre = document.createElement('pre');
-  sqlPre.style.cssText = `
-    margin: 0;
-    padding: 12px;
-    font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-    font-size: 12px;
-    line-height: 1.5;
-    white-space: pre-wrap;
-    word-wrap: break-word;
-    color: #24292e;
-  `;
-
-  // Format SQL first, then apply syntax highlighting
-  const formattedSQL = formatSQLQuery(sqlText);
-  sqlPre.innerHTML = highlightSQLBasic(formattedSQL);
-
-  sqlContainer.appendChild(sqlPre);
-  globalTooltip.appendChild(title);
-  globalTooltip.appendChild(sqlContainer);
-  document.body.appendChild(globalTooltip);
-
-  // Position tooltip with better spacing for easier mouse access
-  const rect = (event.target as HTMLElement).getBoundingClientRect();
-  const tooltipRect = globalTooltip.getBoundingClientRect();
-
-  // Position tooltip to the right of the menu item with some overlap for easier access
-  let left = rect.right - 10; // Small overlap for easier mouse movement
-  let top = rect.top - 5; // Slight vertical offset for better positioning
-
-  // Adjust if tooltip would go off screen
-  if (left + tooltipRect.width > window.innerWidth) {
-    left = rect.left - tooltipRect.width + 10; // Position tooltip on the left with overlap
-  }
-  if (top + tooltipRect.height > window.innerHeight) {
-    top = window.innerHeight - tooltipRect.height - 10;
-  }
-  if (top < 10) {
-    top = 10; // Ensure tooltip doesn't go above viewport
-  }
-
-  globalTooltip.style.left = `${left}px`;
-  globalTooltip.style.top = `${top}px`;
-
-  // Add hover listeners to keep tooltip visible when mouse enters it
-  globalTooltip.addEventListener('mouseenter', () => {
-    // Clear any pending hide timeout when mouse enters tooltip
-    if (tooltipHideTimeout) {
-      clearTimeout(tooltipHideTimeout);
-      tooltipHideTimeout = null;
-    }
-  });
-
-  globalTooltip.addEventListener('mouseleave', () => {
-    // Clear any existing hide timeout
-    if (tooltipHideTimeout) {
-      clearTimeout(tooltipHideTimeout);
-    }
-    // Add a small delay before hiding to prevent accidental hiding
-    tooltipHideTimeout = window.setTimeout(() => {
-      hideSQLTooltip();
-      if (resetHoveredJobref) {
-        resetHoveredJobref();
-      }
-    }, 100);
-  });
-}
-
-/**
- * Hide SQL tooltip
- */
-function hideSQLTooltip(): void {
-  // Clear any pending timeouts
-  if (tooltipHideTimeout) {
-    clearTimeout(tooltipHideTimeout);
-    tooltipHideTimeout = null;
-  }
-  if (tooltipShowTimeout) {
-    clearTimeout(tooltipShowTimeout);
-    tooltipShowTimeout = null;
-  }
-
-  if (globalTooltip) {
-    globalTooltip.remove();
-    globalTooltip = null;
-  }
-}
-
-/**
  * Create a beautiful SQL query card for display
+ * @deprecated Use SQLHoverTooltip.createSQLCard instead
  */
 export function createSQLCard(sqlQuery: string, title?: string): HTMLElement {
-  const card = document.createElement('div');
-  card.className = 'sql-card';
-  card.style.cssText = `
-    background: #ffffff;
-    border: 1px solid #e1e4e8;
-    border-radius: 8px;
-    padding: 16px;
-    margin: 8px 0;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-    font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-    font-size: 13px;
-    line-height: 1.5;
-  `;
-
-  if (title) {
-    const titleEl = document.createElement('div');
-    titleEl.className = 'sql-card-title';
-    titleEl.style.cssText = `
-      font-weight: 600;
-      color: #24292e;
-      margin-bottom: 8px;
-      font-size: 14px;
-    `;
-    titleEl.textContent = title;
-    card.appendChild(titleEl);
-  }
-
-  const sqlEl = document.createElement('pre');
-  sqlEl.className = 'sql-card-content';
-  sqlEl.style.cssText = `
-    margin: 0;
-    white-space: pre-wrap;
-    word-wrap: break-word;
-    color: #24292e;
-    background: #f6f8fa;
-    padding: 12px;
-    border-radius: 4px;
-    border: 1px solid #e1e4e8;
-  `;
-
-  // Format SQL first, then apply syntax highlighting
-  const formattedSQL = formatSQLQuery(sqlQuery);
-  sqlEl.innerHTML = highlightSQLBasic(formattedSQL);
-  card.appendChild(sqlEl);
-
-  return card;
-}
-
-/**
- * Format SQL query with standard formatting rules
- */
-function formatSQLQuery(sql: string): string {
-  try {
-    return formatSQL(sql, {
-      language: 'sql',
-      tabWidth: 2,
-      keywordCase: 'upper',
-      dataTypeCase: 'upper',
-      functionCase: 'upper',
-      identifierCase: 'preserve',
-      indentStyle: 'standard',
-      logicalOperatorNewline: 'before',
-      expressionWidth: 50,
-      linesBetweenQueries: 2
-    });
-  } catch (error) {
-    // If formatting fails, return original SQL
-    console.warn('SQL formatting failed:', error);
-    return sql;
-  }
-}
-
-/**
- * Basic SQL syntax highlighting function
- */
-function highlightSQLBasic(sql: string): string {
-  // SQL keywords to highlight
-  const keywords = [
-    'SELECT',
-    'FROM',
-    'WHERE',
-    'ORDER',
-    'BY',
-    'GROUP',
-    'HAVING',
-    'JOIN',
-    'INNER',
-    'LEFT',
-    'RIGHT',
-    'OUTER',
-    'ON',
-    'AS',
-    'AND',
-    'OR',
-    'NOT',
-    'IN',
-    'EXISTS',
-    'BETWEEN',
-    'LIKE',
-    'IS',
-    'NULL',
-    'DISTINCT',
-    'LIMIT',
-    'INSERT',
-    'UPDATE',
-    'DELETE',
-    'CREATE',
-    'DROP',
-    'ALTER',
-    'TABLE',
-    'INDEX',
-    'VIEW',
-    'PROCEDURE',
-    'FUNCTION',
-    'TRIGGER',
-    'DATABASE',
-    'SCHEMA',
-    'UNION',
-    'ALL',
-    'CASE',
-    'WHEN',
-    'THEN',
-    'ELSE',
-    'END',
-    'IF',
-    'WHILE',
-    'FOR',
-    'LOOP',
-    'BEGIN',
-    'COMMIT',
-    'ROLLBACK',
-    'TRANSACTION',
-    'GRANT',
-    'REVOKE',
-    'PRIMARY',
-    'KEY',
-    'FOREIGN',
-    'REFERENCES',
-    'CONSTRAINT',
-    'CHECK',
-    'DEFAULT',
-    'AUTO_INCREMENT',
-    'VARCHAR',
-    'INT',
-    'BIGINT',
-    'SMALLINT',
-    'TINYINT',
-    'DECIMAL',
-    'FLOAT',
-    'DOUBLE',
-    'CHAR',
-    'TEXT',
-    'DATE',
-    'TIME',
-    'DATETIME',
-    'TIMESTAMP',
-    'BOOLEAN',
-    'BLOB',
-    'JSON'
-  ];
-
-  // Escape HTML entities first to prevent double-escaping
-  let highlighted = sql
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#39;');
-
-  // Highlight keywords
-  keywords.forEach(keyword => {
-    const regex = new RegExp(`\\b${keyword}\\b`, 'gi');
-    highlighted = highlighted.replace(
-      regex,
-      `<span style="color: #cf222e; font-weight: bold;">${keyword.toUpperCase()}</span>`
-    );
-  });
-
-  // Highlight strings (single and double quotes)
-  highlighted = highlighted.replace(
-    /(&#39;)((?:\\.|(?!\1)[^\\])*?)\1/g,
-    '<span style="color: #0a3069;">$1$2$1</span>'
-  );
-  highlighted = highlighted.replace(
-    /(&quot;)((?:\\.|(?!\1)[^\\])*?)\1/g,
-    '<span style="color: #0a3069;">$1$2$1</span>'
-  );
-
-  // Highlight numbers
-  highlighted = highlighted.replace(
-    /\b\d+(\.\d+)?\b/g,
-    '<span style="color: #0a3069; font-weight: bold;">$&</span>'
-  );
-
-  // Highlight comments (-- and /* */)
-  highlighted = highlighted.replace(
-    /--.*$/gm,
-    '<span style="color: #6a737d; font-style: italic;">$&</span>'
-  );
-  highlighted = highlighted.replace(
-    /\/\*[\s\S]*?\*\//g,
-    '<span style="color: #6a737d; font-style: italic;">$&</span>'
-  );
-
-  return highlighted;
+  return SQLHoverTooltip.createSQLCard(sqlQuery, title);
 }
 
 /**
